Browse the metadata of projects contained in the popler database

pplr_browse() reports the metadata of LTER studies contained in the popler database. The user can subset which datasets, and which metadata variables to visualize.

Usage

pplr_browse(..., full_tbl = FALSE, vars = NULL, view = FALSE,
  keyword = NULL, report = FALSE)

Arguments

...

A logical expression to subset the table containing the metadata of datasets contained in popler

full_tbl

logical; Should the function returns the standard columns, or the full main table? Default is FALSE.

vars

A vector of characters in case the user want to select which variables of popler's main table should be selected?

view

If TRUE, opens up a spreadsheet-style data viewer.

keyword

A string used to select individual datasets based on pattern matching. The string is matched to every string element in the variables of the metadata table in popler.

report

logical; If TRUE, function produces a markdown report about each study's metadata, and opens it as a html page. Default is FALSE.

Value

A data frame combining the metadata of each project and the taxonomic units associated with each project.

This data frame is of class popler, data.frame, tbl_df, and tbl.

Examples

if (FALSE) {
# No arguments return the standard 16 columns of popler's main table
default_vars = pplr_browse()

# full_tbl = TRUE returns the full table
all_vars = pplr_browse(full_tbl = TRUE)

# subset only data from the sevilleta LTER, and open the relative report in a html page
sev_data = pplr_browse(lterid == "SEV", report = TRUE)

# consider only plant data sets 
plant_data = pplr_browse(kingdom == "Plantae")

# Select only the data you need
three_columns = pplr_browse(vars = c("title","proj_metadata_key","genus","species"))

# Select only the data you need
study_21 = pplr_browse( proj_metadata_key == 25)

# Select studies that contain word "parasite"
parasite_studies = pplr_browse( keyword = "parasite")
}