Download data from the popler database

This function downloads datasets contained in the popler database. The user can download data directly, using a logical expression, or indirectly, using objects created by pplr_browse.

Usage

pplr_get_data(..., cov_unpack = FALSE)

Arguments

...

An object produced by pplr_browse or a logical expression.

cov_unpack

logical; if TRUE, function pplr_cov_unpack is applied to the variable covariates of the downloaded dataset in order to extract the variables contained in therein and combine the new columns with the default output. Default is FALSE.

Value

This data fame is of class get_data, and data.frame.

Details

. By default, the following variables are included when a user calls pplr_get_data().

authors

authors_contact

year

day

month

sppcode

genus

species

datatype

spatial_replication_level_1_label

spatial_replication_level_1

spatial_replication_level_2_label

spatial_replication_level_2

spatial_replication_level_3_label

spatial_replication_level_3

spatial_replication_level_4_label

spatial_replication_level_4

spatial_replication_level_5_label

spatial_replication_level_5

proj_metadata_key

structure_type_1

structure_type_2

structure_type_3

structure_type_4

treatment_type_1

treatment_type_2

treatment_type_3

covariates

Examples

if (FALSE) {
# browse a study, then get the data associated with it
parasite = pplr_browse(proj_metadata_key == 25)
gh_data = pplr_get_data(parasite)

# insect data sets from the SEV lter site
insect_sev = pplr_browse(class == "Insecta" & lterid == "SEV")
insect_25_yrs96_99 = pplr_get_data(insect_sev)

insect_21_25 = pplr_get_data( (proj_metadata_key == 43 | 
                               proj_metadata_key == 25) )
}